server/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
backports.functools-lru-cache==2.0.0
bidict==0.23.1
blinker==1.9.0
boto3==1.38.36
botocore==1.38.36
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
distro==1.9.0
dnspython==2.7.0
eventlet==0.40.0
fastapi==0.115.12
ffmpeg-python==0.2.0
filelock==3.18.0
Flask==3.1.1
flask-cors==6.0.1
Flask-SocketIO==5.5.1
fsspec==2025.5.1
future==1.0.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.0
google-api-python-client==2.172.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-speech==2.33.0
google-cloud-translate==3.20.3
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
greenlet==3.2.3
grpc-google-iam-v1==0.14.2
grpcio==1.73.0
grpcio-status==1.71.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.32.6
hyperframe==6.1.0
idna==3.10
indic-transliteration==2.3.69
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
joblib==1.5.1
llvmlite==0.44.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
more-itertools==10.7.0
mpmath==1.3.0
networkx==3.5
nltk==3.9.1
numba==0.61.2
numpy==2.2.6
openai==1.86.0
openai-whisper==20240930
packaging==25.0
pandas==2.3.0
pillow==11.2.1
portalocker==2.10.1
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-engineio==4.12.2
python-socketio==5.13.0
pytz==2025.2
pywin32==310
PyYAML==6.0.2
qdrant-client==1.14.2
regex==2024.11.6
requests==2.32.4
rich==14.0.0
roman==5.0
rsa==4.9.1
s3transfer==0.13.0
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.15.3
sentence-transformers==4.1.0
setuptools==80.9.0
shellingham==1.5.4
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
sounddevice==0.5.2
soundfile==0.13.1
starlette==0.46.2
sympy==1.13.1
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.1
toml==0.10.2
torch==2.5.1+cu121
torchaudio==2.5.1+cu121
torchvision==0.20.1+cu121
tqdm==4.67.1
transformers==4.52.4
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.4.0
uvicorn==0.34.3
websocket-client==1.8.0
Werkzeug==3.1.3
wsproto==1.2.0
